Posts and comments combined and manual data
Comments and posts are now under the same table, "Posts", and can have
parents, such that a comment's parent is the post/comment it is replying
to and posts have a null parent.
Manually added 2 users and 3 posts, one of which is a comment. Note the
slightly different mysql input into Posts command depending on whether
the insert is a comment or not.