Work on the add to Posts action path
Continuing to flesh out the method by which we add a post to the
database.
The variables "flags" and "rating" now default to 0.
We now have an addPost.php controller that sends along only "title",
"content", and "userName" since all other variables have defaults. THIS
CONTROLLER REQUIRES POST AND SESSION DATA WHICH DO NOT YET EXIST.
The add function in the model posts.php is no longer an artifact and now
adds data correctly to the database. STILL HAS NOT BEEN TESTED.