Controllers and functions and JS, oh my!
Made a controller to handle post number increments, such as the rating
number and flag number. posts.php has new functions that the controller
uses. Started an increment.js file to increment numbers without
refreshing the page. Note that the number is not yet displayed on the
page, so increment is not yet possible.