Began on admin profiles
Edited resetdb to have an admin user. Changed a controller and tweaked
one of the models to allow for admin status to be stored in session.
This allows us to do the same thing we did with delete for user posts
only with remove flags, delete any post, as well as grant admin status
to a specified user.