Work on limiting increments
Big update here, but nothing works without the others so can't split it
up.
resetdb.txt now has a rating column in the Post_Ratings table.
posts.php now has functions that correctly change values in the table.
incrementController.php correctly calls those functions.
increment.js correctly sends info to incrementController.php via post
and changes increment numbers on post-display.php
post-display.php now grays out buttons if user is not logged in or if
the resetdb.txt table shows the user as already having hit the button.

NOTE: this commit has a bug! Pressing the flag button brings the user to
the create post page. I think because it's in the form that goes to
postController.php. Needs to be worked on.